基础库屏幕适配问题

+++ plugins/base_framework/lib/extension/size_extension.dart
@@ -3,11 +3,11 @@ import '../base_framework.dart';
 /// 屏幕适配封装
 extension SizeExtension on num {
   ///[ScreenUtil.getWidth]
-  num get w => ScreenUtil().getWidth(this);
+  num get w => ScreenUtil().getWidth(this.toDouble());
 
   ///[ScreenUtil.getHeight]
-  num get h => ScreenUtil().getHeight(this);
+  num get h => ScreenUtil().getHeight(this.toDouble());
 
   ///[ScreenUtil.getSp]
-  num get sp => ScreenUtil().getSp(this);
+  num get sp => ScreenUtil().getSp(this.toDouble());
 }
